slack-bot/requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.23.0
openai>=1.12.0
python-dotenv>=1.0.0
gitpython>=3.1.40
chromadb>=0.4.22
sentence-transformers>=2.3.1
PyYAML>=6.0.1
